Fix up & repot tests
