+++ frame/democracy/src/tests.rs
@@ -25,11 +25,22 @@ use frame_support::{
 };
 use sp_core::H256;
 use sp_runtime::{
-	traits::{BlakeTwo256, IdentityLookup, Bounded, BadOrigin},
+	traits::{BlakeTwo256, IdentityLookup, BadOrigin},
 	testing::Header, Perbill,
 };
 use pallet_balances::{BalanceLock, Error as BalancesError};
 use frame_system::EnsureSignedBy;
+
+mod cancellation;
+mod delegation;
+mod external_proposing;
+mod fast_tracking;
+mod lock_voting;
+mod preimage;
+mod proxying;
+mod public_proposals;
+mod scheduling;
+mod voting;
 
 const AYE: Vote = Vote { aye: true, conviction: Conviction::None };
 const NAY: Vote = Vote { aye: false, conviction: Conviction::None };
@@ -227,1238 +238,6 @@ fn big_nay(who: u64) -> AccountVote<u64> {
 	AccountVote::Standard { vote: BIG_NAY, balance: Balances::free_balance(&who) }
 }
 
-#[test]
-fn missing_preimage_should_fail() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 0);
-	});
+fn tally(r: ReferendumIndex) -> Tally<u64> {
+	Democracy::referendum_status(r).unwrap().tally
 }
-
-#[test]
-fn preimage_deposit_should_be_required_and_returned() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		// fee of 100 is too much.
-		PREIMAGE_BYTE_DEPOSIT.with(|v| *v.borrow_mut() = 100);
-		assert_noop!(
-			Democracy::note_preimage(Origin::signed(6), vec![0; 500]),
-			BalancesError::<Test, _>::InsufficientBalance,
-		);
-		// fee of 1 is reasonable.
-		PREIMAGE_BYTE_DEPOSIT.with(|v| *v.borrow_mut() = 1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		assert_eq!(Balances::reserved_balance(6), 12);
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::reserved_balance(6), 0);
-		assert_eq!(Balances::free_balance(6), 60);
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn preimage_deposit_should_be_reapable_earlier_by_owner() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		PREIMAGE_BYTE_DEPOSIT.with(|v| *v.borrow_mut() = 1);
-		assert_ok!(Democracy::note_preimage(Origin::signed(6), set_balance_proposal(2)));
-
-		assert_eq!(Balances::reserved_balance(6), 12);
-
-		next_block();
-		assert_noop!(
-			Democracy::reap_preimage(Origin::signed(6), set_balance_proposal_hash(2)),
-			Error::<Test>::TooEarly
-		);
-		next_block();
-		assert_ok!(Democracy::reap_preimage(Origin::signed(6), set_balance_proposal_hash(2)));
-
-		assert_eq!(Balances::free_balance(6), 60);
-		assert_eq!(Balances::reserved_balance(6), 0);
-	});
-}
-
-#[test]
-fn preimage_deposit_should_be_reapable() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_noop!(
-			Democracy::reap_preimage(Origin::signed(5), set_balance_proposal_hash(2)),
-			Error::<Test>::PreimageMissing
-		);
-
-		PREIMAGE_BYTE_DEPOSIT.with(|v| *v.borrow_mut() = 1);
-		assert_ok!(Democracy::note_preimage(Origin::signed(6), set_balance_proposal(2)));
-		assert_eq!(Balances::reserved_balance(6), 12);
-
-		next_block();
-		next_block();
-		next_block();
-		assert_noop!(
-			Democracy::reap_preimage(Origin::signed(5), set_balance_proposal_hash(2)),
-			Error::<Test>::TooEarly
-		);
-
-		next_block();
-		assert_ok!(Democracy::reap_preimage(Origin::signed(5), set_balance_proposal_hash(2)));
-		assert_eq!(Balances::reserved_balance(6), 0);
-		assert_eq!(Balances::free_balance(6), 48);
-		assert_eq!(Balances::free_balance(5), 62);
-	});
-}
-
-#[test]
-fn noting_imminent_preimage_for_free_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		PREIMAGE_BYTE_DEPOSIT.with(|v| *v.borrow_mut() = 1);
-
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash(2),
-			VoteThreshold::SuperMajorityApprove,
-			1
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		assert_noop!(
-			Democracy::note_imminent_preimage(Origin::signed(7), set_balance_proposal(2)),
-			Error::<Test>::NotImminent
-		);
-
-		next_block();
-
-		// Now we're in the dispatch queue it's all good.
-		assert_ok!(Democracy::note_imminent_preimage(Origin::signed(7), set_balance_proposal(2)));
-
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn reaping_imminent_preimage_should_fail() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let h = set_balance_proposal_hash_and_note(2);
-		let r = Democracy::inject_referendum(3, h, VoteThreshold::SuperMajorityApprove, 1);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		next_block();
-		next_block();
-		// now imminent.
-		assert_noop!(Democracy::reap_preimage(Origin::signed(6), h), Error::<Test>::Imminent);
-	});
-}
-
-#[test]
-fn external_and_public_interleaving_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(1),
-		));
-		assert_ok!(propose_set_balance_and_note(6, 2, 2));
-
-		fast_forward_to(2);
-
-		// both waiting: external goes first.
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 4,
-				proposal_hash: set_balance_proposal_hash_and_note(1),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-		// replenish external
-		assert_ok!(Democracy::external_propose(
-				Origin::signed(2),
-				set_balance_proposal_hash_and_note(3),
-			));
-
-		fast_forward_to(4);
-
-		// both waiting: public goes next.
-		assert_eq!(
-			Democracy::referendum_status(1),
-			Ok(ReferendumStatus {
-				end: 6,
-				proposal_hash: set_balance_proposal_hash_and_note(2),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-		// don't replenish public
-
-		fast_forward_to(6);
-
-		// it's external "turn" again, though since public is empty that doesn't really matter
-		assert_eq!(
-			Democracy::referendum_status(2),
-			Ok(ReferendumStatus {
-				end: 8,
-				proposal_hash: set_balance_proposal_hash_and_note(3),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-		// replenish external
-		assert_ok!(Democracy::external_propose(
-				Origin::signed(2),
-				set_balance_proposal_hash_and_note(5),
-			));
-
-		fast_forward_to(8);
-
-		// external goes again because there's no public waiting.
-		assert_eq!(
-			Democracy::referendum_status(3),
-			Ok(ReferendumStatus {
-				end: 10,
-				proposal_hash: set_balance_proposal_hash_and_note(5),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-		// replenish both
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(7),
-		));
-		assert_ok!(propose_set_balance_and_note(6, 4, 2));
-
-		fast_forward_to(10);
-
-		// public goes now since external went last time.
-		assert_eq!(
-			Democracy::referendum_status(4),
-			Ok(ReferendumStatus {
-				end: 12,
-				proposal_hash: set_balance_proposal_hash_and_note(4),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-		// replenish public again
-		assert_ok!(propose_set_balance_and_note(6, 6, 2));
-		// cancel external
-		let h = set_balance_proposal_hash_and_note(7);
-		assert_ok!(Democracy::veto_external(Origin::signed(3), h));
-
-		fast_forward_to(12);
-
-		// public goes again now since there's no external waiting.
-		assert_eq!(
-			Democracy::referendum_status(5),
-			Ok(ReferendumStatus {
-				end: 14,
-				proposal_hash: set_balance_proposal_hash_and_note(6),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-	});
-}
-
-#[test]
-fn emergency_cancel_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			2
-		);
-		assert!(Democracy::referendum_status(r).is_ok());
-
-		assert_noop!(Democracy::emergency_cancel(Origin::signed(3), r), BadOrigin);
-		assert_ok!(Democracy::emergency_cancel(Origin::signed(4), r));
-		assert!(Democracy::referendum_info(r).is_none());
-
-		// some time later...
-
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			2
-		);
-		assert!(Democracy::referendum_status(r).is_ok());
-		assert_noop!(
-			Democracy::emergency_cancel(Origin::signed(4), r),
-			Error::<Test>::AlreadyCanceled,
-		);
-	});
-}
-
-#[test]
-fn veto_external_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(2),
-		));
-		assert!(<NextExternal<Test>>::exists());
-
-		let h = set_balance_proposal_hash_and_note(2);
-		assert_ok!(Democracy::veto_external(Origin::signed(3), h.clone()));
-		// cancelled.
-		assert!(!<NextExternal<Test>>::exists());
-		// fails - same proposal can't be resubmitted.
-		assert_noop!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash(2),
-		), Error::<Test>::ProposalBlacklisted);
-
-		fast_forward_to(1);
-		// fails as we're still in cooloff period.
-		assert_noop!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash(2),
-		), Error::<Test>::ProposalBlacklisted);
-
-		fast_forward_to(2);
-		// works; as we're out of the cooloff period.
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(2),
-		));
-		assert!(<NextExternal<Test>>::exists());
-
-		// 3 can't veto the same thing twice.
-		assert_noop!(
-			Democracy::veto_external(Origin::signed(3), h.clone()),
-			Error::<Test>::AlreadyVetoed
-		);
-
-		// 4 vetoes.
-		assert_ok!(Democracy::veto_external(Origin::signed(4), h.clone()));
-		// cancelled again.
-		assert!(!<NextExternal<Test>>::exists());
-
-		fast_forward_to(3);
-		// same proposal fails as we're still in cooloff
-		assert_noop!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash(2),
-		), Error::<Test>::ProposalBlacklisted);
-		// different proposal works fine.
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(3),
-		));
-	});
-}
-
-#[test]
-fn external_referendum_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_noop!(
-			Democracy::external_propose(
-				Origin::signed(1),
-				set_balance_proposal_hash(2),
-			),
-			BadOrigin,
-		);
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(2),
-		));
-		assert_noop!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash(1),
-		), Error::<Test>::DuplicateProposal);
-		fast_forward_to(2);
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 4,
-				proposal_hash: set_balance_proposal_hash(2),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-	});
-}
-
-#[test]
-fn external_majority_referendum_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_noop!(
-			Democracy::external_propose_majority(
-				Origin::signed(1),
-				set_balance_proposal_hash(2)
-			),
-			BadOrigin,
-		);
-		assert_ok!(Democracy::external_propose_majority(
-			Origin::signed(3),
-			set_balance_proposal_hash_and_note(2)
-		));
-		fast_forward_to(2);
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 4,
-				proposal_hash: set_balance_proposal_hash(2),
-				threshold: VoteThreshold::SimpleMajority,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-	});
-}
-
-#[test]
-fn external_default_referendum_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_noop!(
-			Democracy::external_propose_default(
-				Origin::signed(3),
-				set_balance_proposal_hash(2)
-			),
-			BadOrigin,
-		);
-		assert_ok!(Democracy::external_propose_default(
-			Origin::signed(1),
-			set_balance_proposal_hash_and_note(2)
-		));
-		fast_forward_to(2);
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 4,
-				proposal_hash: set_balance_proposal_hash(2),
-				threshold: VoteThreshold::SuperMajorityAgainst,
-				delay: 2,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-	});
-}
-
-#[test]
-fn fast_track_referendum_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		let h = set_balance_proposal_hash_and_note(2);
-		assert_noop!(Democracy::fast_track(Origin::signed(5), h, 3, 2), Error::<Test>::ProposalMissing);
-		assert_ok!(Democracy::external_propose_majority(
-			Origin::signed(3),
-			set_balance_proposal_hash_and_note(2)
-		));
-		assert_noop!(Democracy::fast_track(Origin::signed(1), h, 3, 2), BadOrigin);
-		assert_ok!(Democracy::fast_track(Origin::signed(5), h, 2, 0));
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 2,
-				proposal_hash: set_balance_proposal_hash_and_note(2),
-				threshold: VoteThreshold::SimpleMajority,
-				delay: 0,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-	});
-}
-
-#[test]
-fn instant_referendum_works() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		let h = set_balance_proposal_hash_and_note(2);
-		assert_noop!(Democracy::fast_track(Origin::signed(5), h, 3, 2), Error::<Test>::ProposalMissing);
-		assert_ok!(Democracy::external_propose_majority(
-			Origin::signed(3),
-			set_balance_proposal_hash_and_note(2)
-		));
-		assert_noop!(Democracy::fast_track(Origin::signed(1), h, 3, 2), BadOrigin);
-		assert_noop!(Democracy::fast_track(Origin::signed(5), h, 1, 0), BadOrigin);
-		assert_noop!(Democracy::fast_track(Origin::signed(6), h, 1, 0), Error::<Test>::InstantNotAllowed);
-		INSTANT_ALLOWED.with(|v| *v.borrow_mut() = true);
-		assert_ok!(Democracy::fast_track(Origin::signed(6), h, 1, 0));
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 1,
-				proposal_hash: set_balance_proposal_hash_and_note(2),
-				threshold: VoteThreshold::SimpleMajority,
-				delay: 0,
-				tally: Tally { ayes: 0, nays: 0, turnout: 0 },
-			})
-		);
-	});
-}
-
-#[test]
-fn fast_track_referendum_fails_when_no_simple_majority() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		let h = set_balance_proposal_hash_and_note(2);
-		assert_ok!(Democracy::external_propose(
-			Origin::signed(2),
-			set_balance_proposal_hash_and_note(2)
-		));
-		assert_noop!(
-			Democracy::fast_track(Origin::signed(5), h, 3, 2),
-			Error::<Test>::NotSimpleMajority
-		);
-	});
-}
-
-#[test]
-fn backing_for_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_ok!(propose_set_balance_and_note(1, 2, 2));
-		assert_ok!(propose_set_balance_and_note(1, 4, 4));
-		assert_ok!(propose_set_balance_and_note(1, 3, 3));
-		assert_eq!(Democracy::backing_for(0), Some(2));
-		assert_eq!(Democracy::backing_for(1), Some(4));
-		assert_eq!(Democracy::backing_for(2), Some(3));
-	});
-}
-
-#[test]
-fn single_proposal_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-		assert!(Democracy::referendum_info(0).is_none());
-
-		// start of 2 => next referendum scheduled.
-		fast_forward_to(2);
-
-		let r = 0;
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		assert_eq!(Democracy::referendum_count(), 1);
-		assert_eq!(
-			Democracy::referendum_status(0),
-			Ok(ReferendumStatus {
-				end: 4,
-				proposal_hash: set_balance_proposal_hash_and_note(2),
-				threshold: VoteThreshold::SuperMajorityApprove,
-				delay: 2,
-				tally: Tally { ayes: 1, nays: 0, turnout: 1 },
-			})
-		);
-
-		fast_forward_to(3);
-
-		// referendum still running
-		assert!(Democracy::referendum_status(0).is_ok());
-
-		// referendum runs during 2 and 3, ends @ start of 4.
-		fast_forward_to(4);
-
-		assert!(Democracy::referendum_status(0).is_err());
-		assert_eq!(Democracy::dispatch_queue(), vec![
-			(6, set_balance_proposal_hash_and_note(2), 0)
-		]);
-
-		// referendum passes and wait another two blocks for enactment.
-		fast_forward_to(6);
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn cancel_queued_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		// start of 2 => next referendum scheduled.
-		fast_forward_to(2);
-
-		assert_ok!(Democracy::vote(Origin::signed(1), 0, aye(1)));
-
-		fast_forward_to(4);
-
-		assert_eq!(Democracy::dispatch_queue(), vec![
-			(6, set_balance_proposal_hash_and_note(2), 0)
-		]);
-
-		assert_noop!(Democracy::cancel_queued(Origin::ROOT, 1), Error::<Test>::ProposalMissing);
-		assert_ok!(Democracy::cancel_queued(Origin::ROOT, 0));
-		assert_eq!(Democracy::dispatch_queue(), vec![]);
-	});
-}
-
-#[test]
-fn proxy_should_work() {
-	new_test_ext().execute_with(|| {
-		assert_eq!(Democracy::proxy(10), None);
-		assert!(System::allow_death(&10));
-
-		assert_noop!(Democracy::activate_proxy(Origin::signed(1), 10), Error::<Test>::NotOpen);
-
-		assert_ok!(Democracy::open_proxy(Origin::signed(10), 1));
-		assert!(!System::allow_death(&10));
-		assert_eq!(Democracy::proxy(10), Some(ProxyState::Open(1)));
-
-		assert_noop!(Democracy::activate_proxy(Origin::signed(2), 10), Error::<Test>::WrongOpen);
-		assert_ok!(Democracy::activate_proxy(Origin::signed(1), 10));
-		assert_eq!(Democracy::proxy(10), Some(ProxyState::Active(1)));
-
-		// Can't set when already set.
-		assert_noop!(Democracy::activate_proxy(Origin::signed(2), 10), Error::<Test>::AlreadyProxy);
-
-		// But this works because 11 isn't proxying.
-		assert_ok!(Democracy::open_proxy(Origin::signed(11), 2));
-		assert_ok!(Democracy::activate_proxy(Origin::signed(2), 11));
-		assert_eq!(Democracy::proxy(10), Some(ProxyState::Active(1)));
-		assert_eq!(Democracy::proxy(11), Some(ProxyState::Active(2)));
-
-		// 2 cannot fire 1's proxy:
-		assert_noop!(Democracy::deactivate_proxy(Origin::signed(2), 10), Error::<Test>::WrongProxy);
-
-		// 1 deactivates their proxy:
-		assert_ok!(Democracy::deactivate_proxy(Origin::signed(1), 10));
-		assert_eq!(Democracy::proxy(10), Some(ProxyState::Open(1)));
-		// but the proxy account cannot be killed until the proxy is closed.
-		assert!(!System::allow_death(&10));
-
-		// and then 10 closes it completely:
-		assert_ok!(Democracy::close_proxy(Origin::signed(10)));
-		assert_eq!(Democracy::proxy(10), None);
-		assert!(System::allow_death(&10));
-
-		// 11 just closes without 2's "permission".
-		assert_ok!(Democracy::close_proxy(Origin::signed(11)));
-		assert_eq!(Democracy::proxy(11), None);
-		assert!(System::allow_death(&11));
-	});
-}
-
-#[test]
-fn single_proposal_should_work_with_proxy() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		fast_forward_to(2);
-		let r = 0;
-		assert_ok!(Democracy::open_proxy(Origin::signed(10), 1));
-		assert_ok!(Democracy::activate_proxy(Origin::signed(1), 10));
-		assert_ok!(Democracy::proxy_vote(Origin::signed(10), r, aye(1)));
-
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 1, nays: 0, turnout: 1 });
-
-		fast_forward_to(6);
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn single_proposal_should_work_with_delegation() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		fast_forward_to(2);
-
-		// Delegate first vote.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::None, 20));
-		let r = 0;
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 3, nays: 0, turnout: 1 });
-
-		// Delegate a second vote.
-		assert_ok!(Democracy::delegate(Origin::signed(3), 1, Conviction::None, 30));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 6, nays: 0, turnout: 1 });
-
-		// Reduce first vote.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::None, 10));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 5, nays: 0, turnout: 1 });
-
-		// Second vote delegates to first; we don't do tiered delegation, so it doesn't get used.
-		assert_ok!(Democracy::delegate(Origin::signed(3), 2, Conviction::None, 30));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 2, nays: 0, turnout: 1 });
-
-		// Main voter cancels their vote
-		assert_ok!(Democracy::unvote(Origin::signed(1), r));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 0, nays: 0, turnout: 0 });
-
-		// First delegator delegates half funds with conviction; nothing changes yet.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::Locked1x, 10));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 0, nays: 0, turnout: 0 });
-
-		// Main voter reinstates their vote
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		assert_eq!(Democracy::referendum_status(r).unwrap().tally, Tally { ayes: 11, nays: 0, turnout: 1 });
-
-		fast_forward_to(6);
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-/*
-#[test]
-fn single_proposal_should_work_with_cyclic_delegation() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		fast_forward_to(2);
-
-		// Check behavior with cycle.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::max_value()));
-		assert_ok!(Democracy::delegate(Origin::signed(3), 2, Conviction::max_value()));
-		assert_ok!(Democracy::delegate(Origin::signed(1), 3, Conviction::max_value()));
-		let r = 0;
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		assert_eq!(Democracy::voters_for(r), vec![1]);
-
-		// Delegated vote is counted.
-		assert_eq!(Democracy::tally(r), (6, 0, 6));
-
-		fast_forward_to(6);
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-/// If transactor already voted, delegated vote is overwritten.
-fn single_proposal_should_work_with_vote_and_delegation() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		fast_forward_to(2);
-
-		let r = 0;
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		// Vote.
-		assert_ok!(Democracy::vote(Origin::signed(2), r, aye(2));
-		// Delegate vote.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::max_value()));
-		assert_eq!(Democracy::voters_for(r), vec![1, 2]);
-		assert_eq!(Democracy::vote_of((r, 1)), AYE);
-		// Delegated vote is not counted.
-		assert_eq!(Democracy::tally(r), (3, 0, 3));
-
-		fast_forward_to(6);
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn single_proposal_should_work_with_undelegation() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		// Delegate and undelegate vote.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::max_value()));
-		assert_ok!(Democracy::undelegate(Origin::signed(2)));
-
-		fast_forward_to(2);
-		let r = 0;
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		assert_eq!(Democracy::referendum_count(), 1);
-		assert_eq!(Democracy::voters_for(r), vec![1]);
-		assert_eq!(Democracy::vote_of((r, 1)), AYE);
-
-		// Delegated vote is not counted.
-		assert_eq!(Democracy::tally(r), (1, 0, 1));
-
-		fast_forward_to(6);
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-/// If transactor voted, delegated vote is overwritten.
-fn single_proposal_should_work_with_delegation_and_vote() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-
-		assert_ok!(propose_set_balance_and_note(1, 2, 1));
-
-		fast_forward_to(2);
-		let r = 0;
-
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		// Delegate vote.
-		assert_ok!(Democracy::delegate(Origin::signed(2), 1, Conviction::max_value()));
-
-		// Vote.
-		assert_ok!(Democracy::vote(Origin::signed(2), r, aye(2));
-
-		assert_eq!(Democracy::referendum_count(), 1);
-		assert_eq!(Democracy::voters_for(r), vec![1, 2]);
-		assert_eq!(Democracy::vote_of((r, 1)), AYE);
-
-		// Delegated vote is not counted.
-		assert_eq!(Democracy::tally(r), (3, 0, 3));
-
-		fast_forward_to(6);
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn deposit_for_proposals_should_be_taken() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_ok!(propose_set_balance_and_note(1, 2, 5));
-		assert_ok!(Democracy::second(Origin::signed(2), 0));
-		assert_ok!(Democracy::second(Origin::signed(5), 0));
-		assert_ok!(Democracy::second(Origin::signed(5), 0));
-		assert_ok!(Democracy::second(Origin::signed(5), 0));
-		assert_eq!(Balances::free_balance(1), 5);
-		assert_eq!(Balances::free_balance(2), 15);
-		assert_eq!(Balances::free_balance(5), 35);
-	});
-}
-
-#[test]
-fn deposit_for_proposals_should_be_returned() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_ok!(propose_set_balance_and_note(1, 2, 5));
-		assert_ok!(Democracy::second(Origin::signed(2), 0));
-		assert_ok!(Democracy::second(Origin::signed(5), 0));
-		assert_ok!(Democracy::second(Origin::signed(5), 0));
-		assert_ok!(Democracy::second(Origin::signed(5), 0));
-		fast_forward_to(3);
-		assert_eq!(Balances::free_balance(1), 10);
-		assert_eq!(Balances::free_balance(2), 20);
-		assert_eq!(Balances::free_balance(5), 50);
-	});
-}
-
-#[test]
-fn proposal_with_deposit_below_minimum_should_not_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_noop!(propose_set_balance(1, 2, 0), Error::<Test>::ValueLow);
-	});
-}
-
-#[test]
-fn poor_proposer_should_not_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_noop!(propose_set_balance(1, 2, 11), BalancesError::<Test, _>::InsufficientBalance);
-	});
-}
-
-#[test]
-fn poor_seconder_should_not_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		assert_ok!(propose_set_balance_and_note(2, 2, 11));
-		assert_noop!(Democracy::second(Origin::signed(1), 0), BalancesError::<Test, _>::InsufficientBalance);
-	});
-}
-
-#[test]
-fn runners_up_should_come_after() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		assert_ok!(propose_set_balance_and_note(1, 2, 2));
-		assert_ok!(propose_set_balance_and_note(1, 4, 4));
-		assert_ok!(propose_set_balance_and_note(1, 3, 3));
-		fast_forward_to(2);
-		assert_ok!(Democracy::vote(Origin::signed(1), 0, aye(1)));
-		fast_forward_to(4);
-		assert_ok!(Democracy::vote(Origin::signed(1), 1, aye(1)));
-		fast_forward_to(6);
-		assert_ok!(Democracy::vote(Origin::signed(1), 2, aye(1)));
-	});
-}
-
-#[test]
-fn ooo_inject_referendums_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r1 = Democracy::inject_referendum(
-			3,
-			set_balance_proposal_hash_and_note(3),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		let r2 = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-
-		assert_ok!(Democracy::vote(Origin::signed(1), r2, aye(1)));
-		assert_eq!(Democracy::voters_for(r2), vec![1]);
-		assert_eq!(Democracy::vote_of((r2, 1)), AYE);
-		assert_eq!(Democracy::tally(r2), (1, 0, 1));
-
-		next_block();
-		assert_eq!(Balances::free_balance(42), 2);
-
-		assert_ok!(Democracy::vote(Origin::signed(1), r1, aye(1)));
-		assert_eq!(Democracy::voters_for(r1), vec![1]);
-		assert_eq!(Democracy::vote_of((r1, 1)), AYE);
-		assert_eq!(Democracy::tally(r1), (1, 0, 1));
-
-		next_block();
-		assert_eq!(Balances::free_balance(42), 3);
-	});
-}
-
-#[test]
-fn simple_passing_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-
-		assert_eq!(Democracy::voters_for(r), vec![1]);
-		assert_eq!(Democracy::vote_of((r, 1)), AYE);
-		assert_eq!(Democracy::tally(r), (1, 0, 1));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn cancel_referendum_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		assert_ok!(Democracy::cancel_referendum(Origin::ROOT, r.into()));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 0);
-	});
-}
-
-#[test]
-fn simple_failing_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, nay(1)));
-
-		assert_eq!(Democracy::voters_for(r), vec![1]);
-		assert_eq!(Democracy::vote_of((r, 1)), NAY);
-		assert_eq!(Democracy::tally(r), (0, 1, 1));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 0);
-	});
-}
-
-#[test]
-fn controversial_voting_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-
-		assert_ok!(Democracy::vote(Origin::signed(1), r, big_aye(1)));
-		assert_ok!(Democracy::vote(Origin::signed(2), r, big_nay(2)));
-		assert_ok!(Democracy::vote(Origin::signed(3), r, big_nay(3)));
-		assert_ok!(Democracy::vote(Origin::signed(4), r, big_aye(4)));
-		assert_ok!(Democracy::vote(Origin::signed(5), r, big_nay(5)));
-		assert_ok!(Democracy::vote(Origin::signed(6), r, big_aye(6)));
-
-		assert_eq!(Democracy::tally(r), (110, 100, 210));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn delayed_enactment_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			1
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, aye(1)));
-		assert_ok!(Democracy::vote(Origin::signed(2), r, aye(2)));
-		assert_ok!(Democracy::vote(Origin::signed(3), r, aye(3)));
-		assert_ok!(Democracy::vote(Origin::signed(4), r, aye(4)));
-		assert_ok!(Democracy::vote(Origin::signed(5), r, aye(5)));
-		assert_ok!(Democracy::vote(Origin::signed(6), r, aye(6)));
-
-		assert_eq!(Democracy::tally(r), (21, 0, 21));
-
-		next_block();
-		assert_eq!(Balances::free_balance(42), 0);
-
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn controversial_low_turnout_voting_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(5), r, big_nay(5)));
-		assert_ok!(Democracy::vote(Origin::signed(6), r, big_aye(6)));
-
-		assert_eq!(Democracy::tally(r), (60, 50, 110));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 0);
-	});
-}
-
-#[test]
-fn passing_low_turnout_voting_should_work() {
-	new_test_ext().execute_with(|| {
-		assert_eq!(Balances::free_balance(42), 0);
-		assert_eq!(Balances::total_issuance(), 210);
-
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(4), r, big_aye(4)));
-		assert_ok!(Democracy::vote(Origin::signed(5), r, big_nay(5)));
-		assert_ok!(Democracy::vote(Origin::signed(6), r, big_aye(6)));
-
-		assert_eq!(Democracy::tally(r), (100, 50, 150));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-
-#[test]
-fn lock_voting_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, Vote {
-			aye: false,
-			conviction: Conviction::Locked5x
-		}));
-		assert_ok!(Democracy::vote(Origin::signed(2), r, Vote {
-			aye: true,
-			conviction: Conviction::Locked4x
-		}));
-		assert_ok!(Democracy::vote(Origin::signed(3), r, Vote {
-			aye: true,
-			conviction: Conviction::Locked3x
-		}));
-		assert_ok!(Democracy::vote(Origin::signed(4), r, Vote {
-			aye: true,
-			conviction: Conviction::Locked2x
-		}));
-		assert_ok!(Democracy::vote(Origin::signed(5), r, Vote {
-			aye: false,
-			conviction: Conviction::Locked1x
-		}));
-
-		assert_eq!(Democracy::tally(r), (250, 100, 150));
-
-		fast_forward_to(2);
-
-		assert_eq!(Balances::locks(1), vec![]);
-		assert_eq!(Balances::locks(2), vec![BalanceLock {
-			id: DEMOCRACY_ID,
-			amount: u64::max_value(),
-			reasons: pallet_balances::Reasons::Misc,
-		}]);
-		assert_eq!(Democracy::locks(2), Some(18));
-		assert_eq!(Balances::locks(3), vec![BalanceLock {
-			id: DEMOCRACY_ID,
-			amount: u64::max_value(),
-			reasons: pallet_balances::Reasons::Misc,
-		}]);
-		assert_eq!(Democracy::locks(3), Some(10));
-		assert_eq!(Balances::locks(4), vec![BalanceLock {
-			id: DEMOCRACY_ID,
-			amount: u64::max_value(),
-			reasons: pallet_balances::Reasons::Misc,
-		}]);
-		assert_eq!(Democracy::locks(4), Some(6));
-		assert_eq!(Balances::locks(5), vec![]);
-
-		assert_eq!(Balances::free_balance(42), 2);
-
-		assert_noop!(Democracy::unlock(Origin::signed(1), 1), Error::<Test>::NotLocked);
-
-		fast_forward_to(5);
-		assert_noop!(Democracy::unlock(Origin::signed(1), 4), Error::<Test>::NotExpired);
-		fast_forward_to(6);
-		assert_ok!(Democracy::unlock(Origin::signed(1), 4));
-		assert_noop!(Democracy::unlock(Origin::signed(1), 4), Error::<Test>::NotLocked);
-
-		fast_forward_to(9);
-		assert_noop!(Democracy::unlock(Origin::signed(1), 3), Error::<Test>::NotExpired);
-		fast_forward_to(10);
-		assert_ok!(Democracy::unlock(Origin::signed(1), 3));
-		assert_noop!(Democracy::unlock(Origin::signed(1), 3), Error::<Test>::NotLocked);
-
-		fast_forward_to(17);
-		assert_noop!(Democracy::unlock(Origin::signed(1), 2), Error::<Test>::NotExpired);
-		fast_forward_to(18);
-		assert_ok!(Democracy::unlock(Origin::signed(1), 2));
-		assert_noop!(Democracy::unlock(Origin::signed(1), 2), Error::<Test>::NotLocked);
-	});
-}
-
-#[test]
-fn no_locks_without_conviction_should_work() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(0);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0,
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, Vote {
-			aye: true,
-			conviction: Conviction::None,
-		}));
-
-		fast_forward_to(2);
-
-		assert_eq!(Balances::free_balance(42), 2);
-		assert_eq!(Balances::locks(1), vec![]);
-	});
-}
-
-#[test]
-fn lock_voting_should_work_with_delegation() {
-	new_test_ext().execute_with(|| {
-		System::set_block_number(1);
-		let r = Democracy::inject_referendum(
-			2,
-			set_balance_proposal_hash_and_note(2),
-			VoteThreshold::SuperMajorityApprove,
-			0
-		);
-		assert_ok!(Democracy::vote(Origin::signed(1), r, Vote {
-			aye: false,
-			conviction: Conviction::Locked5x
-		}));
-		assert_ok!(Democracy::vote(Origin::signed(2), r, Vote {
-			aye: true,
-			conviction: Conviction::Locked4x
-		}));
-		assert_ok!(Democracy::vote(Origin::signed(3), r, Vote {
-			aye: true,
-			conviction: Conviction::Locked3x
-		}));
-		assert_ok!(Democracy::delegate(Origin::signed(4), 2, Conviction::Locked2x));
-		assert_ok!(Democracy::vote(Origin::signed(5), r, Vote {
-			aye: false,
-			conviction: Conviction::Locked1x
-		}));
-
-		assert_eq!(Democracy::tally(r), (250, 100, 150));
-
-		next_block();
-		next_block();
-
-		assert_eq!(Balances::free_balance(42), 2);
-	});
-}
-*/
